refactor: separate unit tests from integration tests in pkg/core and pkg/types
- Move all filesystem-dependent tests to *_integration_test.go files
- Keep only pure logic tests in regular *_test.go files
- pkg/core changes:
  - Split direct_executor_validation_test.go (kept isPathWithin unit test)
  - Split templates_test.go (kept GetCompletePackTemplate unit test)
  - Moved entire files that were purely integration tests
  - Consolidated dodotignore tests into existing integration file
- pkg/types changes:
  - Moved action_status_test.go entirely (all filesystem operations)
  - Split pack_test.go (kept GetFilePath unit test)

This separation improves:
- Test clarity: obvious which tests require filesystem
- Test speed: unit tests run without I/O overhead
- Test reliability: unit tests have no external dependencies
- Architecture validation: enforces pure function design
