+++ pkg/core/actions_integration_test.go
