+++ pkg/core/direct_executor_integration_test.go
