+++ pkg/core/display_data_integration_test.go
