+++ pkg/core/pack_helpers_integration_test.go
