+++ pkg/core/pack_status_integration_test.go
