+++ pkg/core/packs_integration_test.go
